requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
langchain==0.3.28
langchain-openai==0.3.35
langchain-community==0.3.27
chromadb==0.5.0
openai==1.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
pandas==2.2.2
requests==2.32.2
tiktoken==0.7.0
